refactor(challenge): extract proof management from driver
