+++ crates/proof/challenge/src/lib.rs
@@ -16,10 +16,13 @@ mod config;
 pub use config::ChallengerConfig;
 
 mod driver;
-pub use driver::{Driver, DriverComponents, DriverConfig, TeeConfig};
+pub use driver::{Driver, DriverComponents};
 
 mod pending;
 pub use pending::{DisputeIntent, PendingProof, PendingProofs, ProofKind, ProofPhase, ProofUpdate};
+
+mod proof_manager;
+pub use proof_manager::DisputeProofManager;
 
 mod proof_adapter;
 pub use proof_adapter::ChallengerProofAdapter;
